Add current sortType hightlight
