+++ src/style/Footer.css
@@ -1,4 +1,4 @@
-footer{
+footer {
   display: flex;
   justify-content: center;
   position: fixed;
@@ -9,27 +9,31 @@ footer{
   border-top: 1px solid #ddd;
   background-color: #fff;
 
-  & .button-wrapper{
+  & .button-wrapper {
     display: inline-block;
-    width: calc((100vw - 20px)/3);
+    width: calc((100vw - 20px) / 3);
 
-    & button{
+    & button {
+      color: #888;
       width: 100%;
       height: 100%;
       background-color: #fff;
       border: none;
-            
-      &:focus{
-        outline: none;
-      }    
 
-      &:active{
+      &:focus {
+        outline: none;
+      }
+
+      &:active {
         background-color: #aaa;
       }
 
-      & svg{
+      & svg {
         display: block;
         margin: 5px auto 0 auto;
+      }
+      &.selected {
+        color: #000;
       }
     }
   }
